Initial Code Commit
This is a script for Photoshop that batch processes on all supported
filetypes (jpg, jpeg, tif, psd) in the chosen directory to be web
ready. The script outputs the new files to the .jpg filetype. The only
saving method is "Save for Web...", there is no print quality exporting
with this script. The accepted filetypes: jpg, jpeg, tif, psd.